A trial at this new inheritance structure. 20% performance regression, in render.
