+++ morph/Visual.h
@@ -90,8 +90,9 @@ namespace morph {
             this->releaseContext();
         }
 
-        void setSwapInterval()
+        void setSwapInterval() final
         {
+            std::cout << "calling glfwSwapInterval (0);\n";
             // Swap as fast as possible (fixes lag of scene with mouse movements)
             glfwSwapInterval (0);
         }
