+++ morph/VisualBase.h
@@ -2,40 +2,54 @@
  * \file
  *
  * Awesome graphics code for high performance graphing and visualisation. This is the
- * base class that sets up GL, leaving choice of window system (GLFW3/Qt/wx/etc) to a
- * derived class such as morph::Visual or morph::qt::viswidget.
- *
- * Normally, a morph::Visual is the *owner* of a GLFW window in which it does its
- * rendering.
- *
- * This is a base class that is ownable, and can be used in other window drawing system
- * such as Qt and wx.
- *
- * Created by Seb James on 2025/03/01, from morph::Visual.h
+ * abstract base class that contains no GL
  *
  * \author Seb James
  * \date March 2025
  */
 #pragma once
 
-#if defined __gl3_h_ || defined __gl_h_ // could get a fuller list from glfw.h
-// GL headers appear to have been externally included.
-#else
-// Include GLAD header
-# define GLAD_GL_IMPLEMENTATION
-#  include <morph/glad/gl.h>
-#endif // GL headers
+#include <morph/gl/version.h>
+#include <morph/VisualModel.h>
+#include <morph/TextFeatures.h>
+#include <morph/TextGeometry.h>
+#include <morph/VisualTextModel.h> // includes VisualResources.h
+#include <morph/VisualCommon.h>
+#include <morph/gl/shaders.h> // for ShaderInfo/LoadShaders
+#include <morph/keys.h>
+#include <morph/version.h>
 
-#include <morph/VisualBase.h>
+#include <morph/VisualResources.h>
+#include <nlohmann/json.hpp>
+#include <morph/CoordArrows.h>
+#include <morph/quaternion.h>
+#include <morph/mat44.h>
+#include <morph/vec.h>
+#include <morph/tools.h>
+
+#include <string>
+#include <array>
+#include <vector>
+#include <memory>
+#include <functional>
+#include <cstddef>
+
+#include <morph/VisualDefaultShaders.h>
+
+// Use Lode Vandevenne's PNG encoder
+#define LODEPNG_NO_COMPILE_DECODER 1
+#define LODEPNG_NO_COMPILE_ANCILLARY_CHUNKS 1
+#include <morph/lodepng.h>
 
 namespace morph {
 
-#ifdef __OSX__
-    // https://stackoverflow.com/questions/35715579/opengl-created-window-size-twice-as-large
-    static constexpr double retinaScale = 2; // deals with quadrant issue on osx
-#else
-    static constexpr double retinaScale = 1; // Qt has devicePixelRatio() to get retinaScale.
-#endif
+    //! Whether to render with perspective or orthographic (or even a cylindrical projection)
+    enum class perspective_type
+    {
+        perspective,
+        orthographic,
+        cylindrical
+    };
 
     /*!
      * VisualOwnable 'scene' base class
@@ -52,7 +66,7 @@ namespace morph {
      * \tparam glver The OpenGL version, encoded as a single int (see morph::gl::version)
      */
     template <int glver = morph::gl::version_4_1>
-    class VisualOwnable : public morph::VisualBase<glver>
+    class VisualBase
     {
     public:
         /*!
@@ -60,38 +74,36 @@ namespace morph {
          * such as a QWidget.  We have to wait on calling init functions until an OpenGL
          * environment is guaranteed to exist.
          */
-        VisualOwnable() { }
+        VisualBase() { }
 
         /*!
          * Construct a new visualiser. The rule is 1 window to one Visual object. So, this creates a
          * new window and a new OpenGL context.
          */
-        VisualOwnable (const int _width, const int _height, const std::string& _title, const bool _version_stdout = true)
+        VisualBase (const int _width, const int _height, const std::string& _title, const bool _version_stdout = true)
             : window_w(_width)
             , window_h(_height)
             , title(_title)
             , version_stdout(_version_stdout)
         {
-            this->init_gl();
+            this->init_gl(); // abstract
         }
 
         //! Deconstruct gl memory/context
-        void deconstructCommon()
-        {
-            if (this->shaders.gprog) {
-                glDeleteProgram (this->shaders.gprog);
-                this->shaders.gprog = 0;
-                this->active_gprog = morph::visgl::graphics_shader_type::none;
-            }
-            if (this->shaders.tprog) {
-                glDeleteProgram (this->shaders.tprog);
-                this->shaders.tprog = 0;
-            }
-            // Free up the Fonts associated with this morph::Visual
-            morph::VisualResources<glver>::i().freetype_deinit (this);
-        }
+        virtual void deconstructCommon() = 0;
 
-        virtual ~VisualOwnable() { this->deconstructCommon(); }
+        //virtual ~VisualBase() { this->deconstructCommon(); }
+
+        // We do not manage OpenGL context, but it is simpler to have a no-op set/releaseContext for some of the GL setup functions
+        virtual void setContext() {}       // no op here
+        virtual void releaseContext() {}   // no op here
+        virtual void setSwapInterval() {}  // no op here
+        virtual void swapBuffers() {}      // no op here
+
+        // A callback friendly wrapper for setContext
+        static void set_context (morph::VisualBase<glver>* _v) { _v->setContext(); };
+        // A callback friendly wrapper for releaseContext
+        static void release_context (morph::VisualBase<glver>* _v) { _v->releaseContext(); };
 
         // Public init that is given a context (window or widget) and then sets up the
         // VisualResource, shaders and so on.
@@ -120,52 +132,9 @@ namespace morph {
             this->freetype_init();
         }
 
-        //! Stores the OpenGL function context version that was loaded
-        int glfn_version = 0;
-
         //! Take a screenshot of the window. Return vec containing width * height or {-1, -1} on
         //! failure. Set transparent_bg to get a transparent background.
-        morph::vec<int, 2> saveImage (const std::string& img_filename, const bool transparent_bg = false)
-        {
-            this->setContext();
-
-            GLint viewport[4]; // current viewport
-            glGetIntegerv (GL_VIEWPORT, viewport);
-
-            morph::vec<int, 2> dims;
-            dims[0] = viewport[2];
-            dims[1] = viewport[3];
-            auto bits = std::make_unique<GLubyte[]>(dims.product() * 4);
-            auto rbits = std::make_unique<GLubyte[]>(dims.product() * 4);
-
-            glFinish(); // finish all commands of OpenGL
-            glPixelStorei (GL_PACK_ALIGNMENT, 1);
-            glPixelStorei (GL_PACK_ROW_LENGTH, 0);
-            glPixelStorei (GL_PACK_SKIP_ROWS, 0);
-            glPixelStorei (GL_PACK_SKIP_PIXELS, 0);
-            glReadPixels (0, 0, dims[0], dims[1], GL_RGBA, GL_UNSIGNED_BYTE, bits.get());
-
-            for (int i = 0; i < dims[1]; ++i) {
-                int rev_line = (dims[1] - i - 1) * 4 * dims[0];
-                int for_line = i * 4 * dims[0];
-                if (transparent_bg) {
-                    for (int j = 0; j < 4 * dims[0]; ++j) {
-                        rbits[rev_line + j] = bits[for_line + j];
-                    }
-                } else {
-                    for (int j = 0; j < 4 * dims[0]; ++j) {
-                        rbits[rev_line + j] = (j % 4 == 3) ? 255 : bits[for_line + j];
-                    }
-                }
-            }
-            unsigned int error = lodepng::encode (img_filename, rbits.get(), dims[0], dims[1]);
-            if (error) {
-                std::cerr << "encoder error " << error << ": " << lodepng_error_text (error) << std::endl;
-                dims.set_from (-1);
-                return dims;
-            }
-            return dims;
-        }
+        virtual morph::vec<int, 2> saveImage (const std::string& img_filename, const bool transparent_bg = false) = 0;
 
         /*!
          * Set up the passed-in VisualModel (or indeed, VisualTextModel) with functions that need access to Visual attributes.
@@ -174,9 +143,9 @@ namespace morph {
         void bindmodel (std::unique_ptr<T>& model)
         {
             model->set_parent (this);
-            model->get_shaderprogs = &morph::VisualOwnable<glver>::get_shaderprogs;
-            model->get_gprog = &morph::VisualOwnable<glver>::get_gprog;
-            model->get_tprog = &morph::VisualOwnable<glver>::get_tprog;
+            model->get_shaderprogs = &morph::VisualBase<glver>::get_shaderprogs;
+            model->get_gprog = &morph::VisualBase<glver>::get_gprog;
+            model->get_tprog = &morph::VisualBase<glver>::get_tprog;
         }
 
         /*!
@@ -301,134 +270,7 @@ namespace morph {
         static void callback_render (morph::VisualOwnable<glver>* _v) { _v->render(); };
 
         //! Render the scene
-        void render() noexcept final
-        {
-            this->setContext();
-
-            if (this->ptype == perspective_type::orthographic || this->ptype == perspective_type::perspective) {
-                if (this->active_gprog != morph::visgl::graphics_shader_type::projection2d) {
-                    if (this->shaders.gprog) { glDeleteProgram (this->shaders.gprog); }
-                    this->shaders.gprog = morph::gl::LoadShaders (this->proj2d_shader_progs);
-                    this->active_gprog = morph::visgl::graphics_shader_type::projection2d;
-                }
-            } else if (this->ptype == perspective_type::cylindrical) {
-                if (this->active_gprog != morph::visgl::graphics_shader_type::cylindrical) {
-                    if (this->shaders.gprog) { glDeleteProgram (this->shaders.gprog); }
-                    this->shaders.gprog = morph::gl::LoadShaders (this->cyl_shader_progs);
-                    this->active_gprog = morph::visgl::graphics_shader_type::cylindrical;
-                }
-            }
-
-            glUseProgram (this->shaders.gprog);
-            glViewport (0, 0, this->window_w * morph::retinaScale, this->window_h * morph::retinaScale);
-
-            // Set the perspective
-            if (this->ptype == perspective_type::orthographic) {
-                this->setOrthographic();
-            } else if (this->ptype == perspective_type::perspective) {
-                this->setPerspective();
-            } else if (this->ptype == perspective_type::cylindrical) {
-                // Set cylindrical-specific uniforms
-                GLint loc_campos = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("cyl_cam_pos"));
-                if (loc_campos != -1) { glUniform4fv (loc_campos, 1, this->cyl_cam_pos.data()); }
-                GLint loc_cyl_radius = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("cyl_radius"));
-                if (loc_cyl_radius != -1) { glUniform1f (loc_cyl_radius, this->cyl_radius); }
-                GLint loc_cyl_height = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("cyl_height"));
-                if (loc_cyl_height != -1) { glUniform1f (loc_cyl_height, this->cyl_height); }
-            } else {
-                // unknown projection
-                return;
-            }
-
-            // Calculate model view transformation - transforming from "model space" to "worldspace".
-            morph::mat44<float> sceneview;
-            if (this->ptype == perspective_type::orthographic || this->ptype == perspective_type::perspective) {
-                // This line translates from model space to world space. Avoid in cyl?
-                sceneview.translate (this->scenetrans); // send backwards into distance
-            }
-            // And this rotation completes the transition from model to world
-            sceneview.rotate (this->rotation);
-
-            // Clear color buffer and **also depth buffer**
-            glClear (GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-            // Set the background colour:
-            glClearBufferfv (GL_COLOR, 0, bgcolour.data());
-
-            // Lighting shader variables
-            //
-            // Ambient light colour
-            GLint loc_lightcol = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("light_colour"));
-            if (loc_lightcol != -1) { glUniform3fv (loc_lightcol, 1, this->light_colour.data()); }
-            // Ambient light intensity
-            GLint loc_ai = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("ambient_intensity"));
-            if (loc_ai != -1) { glUniform1f (loc_ai, this->ambient_intensity); }
-            // Diffuse light position
-            GLint loc_dp = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("diffuse_position"));
-            if (loc_dp != -1) { glUniform3fv (loc_dp, 1, this->diffuse_position.data()); }
-            // Diffuse light intensity
-            GLint loc_di = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("diffuse_intensity"));
-            if (loc_di != -1) { glUniform1f (loc_di, this->diffuse_intensity); }
-
-            // Switch to text shader program and set the projection matrix
-            glUseProgram (this->shaders.tprog);
-            GLint loc_p = glGetUniformLocation (this->shaders.tprog, static_cast<const GLchar*>("p_matrix"));
-            if (loc_p != -1) { glUniformMatrix4fv (loc_p, 1, GL_FALSE, this->projection.mat.data()); }
-
-            // Switch back to the regular shader prog and render the VisualModels.
-            glUseProgram (this->shaders.gprog);
-
-            // Set the projection matrix just once
-            loc_p = glGetUniformLocation (this->shaders.gprog, static_cast<const GLchar*>("p_matrix"));
-            if (loc_p != -1) { glUniformMatrix4fv (loc_p, 1, GL_FALSE, this->projection.mat.data()); }
-
-            if ((this->ptype == perspective_type::orthographic || this->ptype == perspective_type::perspective)
-                && this->showCoordArrows == true) {
-                // Ensure coordarrows centre sphere will be visible on BG:
-                this->coordArrows->setColourForBackground (this->bgcolour); // releases context...
-                this->setContext(); // ...so re-acquire if we're managing it
-
-                if (this->coordArrowsInScene == true) {
-                    this->coordArrows->setSceneMatrix (sceneview);
-                } else {
-                    this->positionCoordArrows();
-                }
-                this->coordArrows->render();
-            }
-
-            morph::mat44<float> scenetransonly;
-            scenetransonly.translate (this->scenetrans);
-
-            auto vmi = this->vm.begin();
-            while (vmi != this->vm.end()) {
-                if ((*vmi)->twodimensional == true) {
-                    // It's a two-d thing. Now what?
-                    (*vmi)->setSceneMatrix (scenetransonly);
-                } else {
-                    (*vmi)->setSceneMatrix (sceneview);
-                }
-                (*vmi)->render();
-                ++vmi;
-            }
-
-            morph::vec<float, 3> v0 = this->textPosition ({-0.8f, 0.8f});
-            if (this->showTitle == true) {
-                // Render the title text
-                this->textModel->setSceneTranslation (v0);
-                this->textModel->setVisibleOn (this->bgcolour);
-                this->textModel->render();
-            }
-
-            auto ti = this->texts.begin();
-            while (ti != this->texts.end()) {
-                (*ti)->setSceneTranslation (v0);
-                (*ti)->setVisibleOn (this->bgcolour);
-                (*ti)->render();
-                ++ti;
-            }
-
-            this->swapBuffers();
-        }
+        virtual void render() noexcept = 0;
 
         //! Compute a translation vector for text position, using Visual::text_z.
         morph::vec<float, 3> textPosition (const morph::vec<float, 2> p0_coord)
@@ -472,8 +314,8 @@ namespace morph {
 
         // These static functions will be set as callbacks in each VisualModel object.
         static morph::visgl::visual_shaderprogs get_shaderprogs (morph::VisualOwnable<glver>* _v) { return _v->shaders; };
-        static GLuint get_gprog (morph::VisualOwnable<glver>* _v) { return _v->shaders.gprog; };
-        static GLuint get_tprog (morph::VisualOwnable<glver>* _v) { return _v->shaders.tprog; };
+        static GLuint get_gprog (morph::VisualBase<glver>* _v) { return _v->shaders.gprog; };
+        static GLuint get_tprog (morph::VisualBase<glver>* _v) { return _v->shaders.tprog; };
 
         //! The colour of ambient and diffuse light sources
         morph::vec<float, 3> light_colour = { 1.0f, 1.0f, 1.0f };
@@ -792,106 +634,11 @@ namespace morph {
         //! ScatterVisual, etc) which are going to be rendered in the scene.
         std::vector<std::unique_ptr<morph::VisualModel<glver>>> vm;
 
-    public:
-#ifdef GLAD_GL // Only define if GL was included with GLAD
-        void init_glad (GLADloadfunc procaddressfn) // need basic version of this in case client code does not use glad
-        {
-            this->glfn_version = gladLoadGL (procaddressfn);
-            if (this->glfn_version == 0) {
-                throw std::runtime_error ("Failed to initialize GLAD GL context");
-            }
-        }
-#endif
-
-    protected:
         // Initialize OpenGL shaders, set some flags (Alpha, Anti-aliasing), read in any external
         // state from json, and set up the coordinate arrows and any VisualTextModels that will be
         // required to render the Visual.
-        void init_gl()
-        {
-            this->setContext(); // if managing context
+        virtual void init_gl() = 0;
 
-            if (this->version_stdout == true) {
-                unsigned char* glv = (unsigned char*)glGetString(GL_VERSION);
-                std::cout << "This is version " << morph::version_string()
-                          << " of morph::Visual<glver=" << morph::gl::version::vstring (glver)
-                          << "> running on OpenGL Version " << glv << std::endl;
-            }
-
-            this->setSwapInterval();
-
-            // Load up the shaders
-            this->proj2d_shader_progs = {
-                {GL_VERTEX_SHADER, "Visual.vert.glsl", morph::getDefaultVtxShader(glver), 0 },
-                {GL_FRAGMENT_SHADER, "Visual.frag.glsl", morph::getDefaultFragShader(glver), 0 }
-            };
-            this->shaders.gprog = morph::gl::LoadShaders (this->proj2d_shader_progs);
-            this->active_gprog = morph::visgl::graphics_shader_type::projection2d;
-
-            // Alternative cylindrical shader for possible later use. (NB: not loaded immediately)
-            this->cyl_shader_progs = {
-                {GL_VERTEX_SHADER, "VisCyl.vert.glsl", morph::getDefaultCylVtxShader(glver), 0 },
-                {GL_FRAGMENT_SHADER, "Visual.frag.glsl", morph::getDefaultFragShader(glver), 0 }
-            };
-
-            // A specific text shader is loaded for text rendering
-            this->text_shader_progs = {
-                {GL_VERTEX_SHADER, "VisText.vert.glsl", morph::getDefaultTextVtxShader(glver), 0 },
-                {GL_FRAGMENT_SHADER, "VisText.frag.glsl" , morph::getDefaultTextFragShader(glver), 0 }
-            };
-            this->shaders.tprog = morph::gl::LoadShaders (this->text_shader_progs);
-
-            // OpenGL options
-            glEnable (GL_DEPTH_TEST);
-            glEnable (GL_BLEND);
-            glBlendFunc (GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-            glDisable (GL_CULL_FACE);
-            morph::gl::Util::checkError (__FILE__, __LINE__);
-
-            // If possible, read in scenetrans and rotation state from a special config file
-            try {
-                nlohmann::json vconf;
-                std::ifstream fi;
-                fi.open ("/tmp/Visual.json", std::ios::in);
-                fi >> vconf;
-                this->scenetrans[0] = vconf.contains("scenetrans_x") ? vconf["scenetrans_x"].get<float>() : this->scenetrans[0];
-                this->scenetrans[1] = vconf.contains("scenetrans_y") ? vconf["scenetrans_y"].get<float>() : this->scenetrans[1];
-                this->scenetrans[2] = vconf.contains("scenetrans_z") ? vconf["scenetrans_z"].get<float>() : this->scenetrans[2];
-                // Place the same numbers into scenetrans_default, too.
-                this->scenetrans_default[0] = this->scenetrans[0];
-                this->scenetrans_default[1] = this->scenetrans[1];
-                this->scenetrans_default[2] = this->scenetrans[2];
-                this->rotation.w = vconf.contains("scenerotn_w") ? vconf["scenerotn_w"].get<float>() : this->rotation.w;
-                this->rotation.x = vconf.contains("scenerotn_x") ? vconf["scenerotn_x"].get<float>() : this->rotation.x;
-                this->rotation.y = vconf.contains("scenerotn_y") ? vconf["scenerotn_y"].get<float>() : this->rotation.y;
-                this->rotation.z = vconf.contains("scenerotn_z") ? vconf["scenerotn_z"].get<float>() : this->rotation.z;
-            } catch (...) {
-                // No problem if we couldn't read /tmp/Visual.json
-            }
-
-            // Use coordArrowsOffset to set the location of the CoordArrows *scene*
-            this->coordArrows = std::make_unique<morph::CoordArrows<glver>>();
-            // For CoordArrows, because we don't add via Visual::addVisualModel(), we
-            // have to set the get_shaderprogs function here:
-            this->bindmodel (this->coordArrows);
-            // And NOW we can proceed to init:
-            this->coordArrows->init (this->coordArrowsLength, this->coordArrowsThickness, this->coordArrowsEm); // sets up text
-            this->coordArrows->finalize(); // VisualModel::finalize releases context (normally this is the right thing)...
-            this->setContext();            // ...but we've got more work to do, so re-acquire context (if we're managing it)
-
-            morph::gl::Util::checkError (__FILE__, __LINE__);
-
-            // Set up the title, which may or may not be rendered
-            morph::TextFeatures title_tf(0.035f, 64);
-            this->textModel = std::make_unique<morph::VisualTextModel<glver>> (title_tf);
-            this->bindmodel (this->textModel);
-            this->textModel->setSceneTranslation ({0.0f, 0.0f, 0.0f});
-            this->textModel->setupText (this->title);
-
-            this->releaseContext();
-        }
-
-    protected:
         //! The window (and OpenGL context) for this Visual
         morph::win_t* window = nullptr;
