+++ morph/qt/viswidget_mx.h
@@ -6,9 +6,9 @@
 struct QOpenGLWidget; // fwd decl
 
 // VisualOwnable is going to be owned by the QOpenGLWidget
-// Define morph::win_t before #including morph/VisualOwnable.h
+// Define morph::win_t before #including morph/VisualOwnableGladMX.h
 namespace morph { using win_t = QOpenGLWidget; }
-#include <morph/VisualOwnable.h>
+#include <morph/VisualOwnableGladMX.h> // Glad multi OpenGL context class
 
 #include <QtWidgets/QOpenGLWidget>
 #include <QOpenGLContext>
@@ -65,9 +65,9 @@ namespace morph {
         template<int widget_index>
         struct viswidget_mx : public QOpenGLWidget //, protected QOpenGLFunctions_4_1_Core
         {
-            // Unlike the GLFW or morph-in-a-QWindow schemes, we hold the morph::VisualOwnable
+            // Unlike the GLFW or morph-in-a-QWindow schemes, we hold the morph::VisualOwnableGladMX
             // inside the widget.
-            morph::VisualOwnable<gl_version> v;
+            morph::VisualOwnableGladMX<gl_version> v;
 
             // In your Qt code, build VisualModels that should be added to the scene and add them to this.
             std::vector<std::unique_ptr<morph::VisualModel<gl_version>>> newvisualmodels;
